+++ RandomSkunk.Results/Operations/AsResult.cs
@@ -7,7 +7,7 @@ public partial struct Maybe<T>
     /// Converts this <see cref="Maybe{T}"/> to an equivalent <see cref="Result{T}"/>. If this is a <c>Success</c> result, then a
     /// <c>Success</c> result with the same value is returned. If this is a <c>Fail</c> result, then a <c>Fail</c> result with
     /// the same error is returned. Otherwise, if this is a <c>None</c> result, then a <c>Fail</c> result with error code
-    /// <see cref="ErrorCodes.NoneResult"/> is returned.
+    /// <see cref="ErrorCodes.NoValue"/> is returned.
     /// </summary>
     /// <returns>The equivalent <see cref="Result{T}"/>.</returns>
     public Result<T> AsResult() =>
@@ -21,7 +21,7 @@ public static partial class ResultExtensions
     /// Converts this <see cref="Maybe{T}"/> to an equivalent <see cref="Result{T}"/>. If this is a <c>Success</c> result, then a
     /// <c>Success</c> result with the same value is returned. If this is a <c>Fail</c> result, then a <c>Fail</c> result with
     /// the same error is returned. Otherwise, if this is a <c>None</c> result, then a <c>Fail</c> result with error code
-    /// <see cref="ErrorCodes.NoneResult"/> is returned.
+    /// <see cref="ErrorCodes.NoValue"/> is returned.
     /// </summary>
     /// <typeparam name="T">The type of the source result value.</typeparam>
     /// <param name="sourceResult">The source result.</param>