+++ RandomSkunk.Results/Operations/Truncate.cs
@@ -19,8 +19,8 @@ public partial struct Maybe<T>
     /// <summary>
     /// Truncates the <see cref="Maybe{T}"/> into an equivalent <see cref="Result"/>. If it is a <c>Success</c> result, then its
     /// value is ignored and a valueless <c>Success</c> result is returned. If it is a <c>None</c> result, then a <c>Fail</c>
-    /// result with error code <see cref="ErrorCodes.NoneResult"/> is returned. Otherwise, a <c>Fail</c> result with the same
-    /// error as is returned.
+    /// result with error code <see cref="ErrorCodes.NoValue"/> is returned. Otherwise, a <c>Fail</c> result with the same error
+    /// as is returned.
     /// </summary>
     /// <returns>An equivalent <see cref="Result"/>.</returns>
     public Result Truncate() =>
@@ -44,8 +44,8 @@ public static partial class ResultExtensions
     /// <summary>
     /// Truncates the <see cref="Maybe{T}"/> into an equivalent <see cref="Result"/>. If it is a <c>Success</c> result, then its
     /// value is ignored and a valueless <c>Success</c> result is returned. If it is a <c>None</c> result, then a <c>Fail</c>
-    /// result with error code <see cref="ErrorCodes.NoneResult"/> is returned. Otherwise, a <c>Fail</c> result with the same
-    /// error as is returned.
+    /// result with error code <see cref="ErrorCodes.NoValue"/> is returned. Otherwise, a <c>Fail</c> result with the same error
+    /// as is returned.
     /// </summary>
     /// <typeparam name="T">The type of the source result value.</typeparam>
     /// <param name="sourceResult">The source result.</param>