+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning].
 
 - RandomSkunk.Results:
     - Change `Maybe.None()` method to property.
+    - Rename `ErrorCodes.NoneResult` to `ErrorCodes.NoValue`.
 
 ### Added